fix: more compilation warnings

+++ Game/Scripts/TutorialSystem.cpp
@@ -73,7 +73,7 @@ void TutorialSystem::Update(float deltaTime)
 void TutorialSystem::TutorialStart()
 {
 	tutorialCurrentState = 0;
-	tutorialTotalStates = tutorialUI.size() - 1;
+	tutorialTotalStates = static_cast<int>(tutorialUI.size()) - 1;
 	currentTutorialUI->Enable();
 	transform2D->SetPosition(initialPos);
 	transform2D->CalculateMatrices();
